# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for UDP IaaS",
        "title": "UDP IaaS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8116",
    "basePath": "/",
    "paths": {
        "/system/info/iso": {
            "get": {
                "description": "Get list of all available ISO images",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ISO"
                ],
                "summary": "Get ISO List",
                "responses": {
                    "200": {
                        "description": "List of ISO images",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/system.ISOInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/system.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vm/create": {
            "post": {
                "description": "Create a new virtual machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VM"
                ],
                "summary": "Create new VM",
                "parameters": [
                    {
                        "description": "VM Creation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vm.CreateVMRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "VM created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "system.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "system.ISOInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "vm.CreateVMRequest": {
            "type": "object",
            "properties": {
                "disk_size_gb": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "iso_path": {
                    "type": "string"
                },
                "memory": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "vcpu": {
                    "type": "integer"
                }
            }
        }
    }
}